requirements.txt
   fastapi
    pydantic
    requests
    pymongo
    python-dotenv